Make light-box reusable
