+++ index.html
@@ -98,27 +98,7 @@
     <!-- back-to-top: start -->
     <a href="#" id="back_to_top">UP</a>
     <!-- back-to-top: end -->
-    <div class="light-box">
-        <div class="light-box-header">
-            <div class="gallery">1/1</div>
-            <div class="icons">
-                <i class="fa fa-square-o" aria-hidden="true"></i>
-                <i class="fa fa-search-minus" aria-hidden="true"></i>
-                <i class="fa fa-search-plus" aria-hidden="true"></i>
-                <i class="fa fa-arrows-alt" aria-hidden="true"></i>
-                <i class="fa fa-download" aria-hidden="true"></i>
-                <i class="fa fa-times" aria-hidden="true"></i>
-            </div>
-        </div>
-        <div class="image">
-            <i class="fa fa-angle-left" aria-hidden="true"></i>
-            <div>
-                <img src="./img/about-3.jpg" alt="about">
-            </div>
-            <i class="fa fa-angle-right" aria-hidden="true"></i>
-        </div>
-        <div class="light-box-footer">Fit image</div>
-    </div>
+    <div class="light-box"></div>
     <script src="./js/main.js" type="module"></script>
 </body>
