+++ js/renderLightBox.js
@@ -3,13 +3,8 @@ function renderLightBox(image, count, size) {
     const ArraySize = document.querySelector('.light-box .gallery');
     const HTML = image;
     const domImage = DOM.innerHTML = HTML;
-
     const text = count + 1 + "/" + size;
     const domSize = ArraySize.innerText = text;
-
-    // return {
-    //     domImage,
-    //     domSize
-    // }
+    return;
 }
 export default renderLightBox;
